+++ app-backend/db/src/main/scala/ProjectDao.scala
@@ -245,6 +245,18 @@ object ProjectDao
     }
   }
 
+  def getShareCount(projectId: UUID, userId: String): ConnectionIO[Long] =
+    ProjectDao
+      .getPermissions(projectId)
+      .map { acrList =>
+        acrList.collect {
+          case Some(ObjectAccessControlRule(subjType, Some(subjectId), _))
+              if subjType == SubjectType.User && subjectId != userId =>
+            subjectId
+        }
+      }
+      .map(_.distinct.length.toLong)
+
   def updateSceneIngestStatus(projectLayerId: UUID): ConnectionIO[Int] = {
     val updateStatusQuery =
       sql"""