+++ app-backend/db/src/test/scala/com/azavea/rf/database/UploadDaoSpec.scala
@@ -39,7 +39,8 @@ class UploadDaoSpec
               datasource <- unsafeGetRandomDatasource
               insertedUpload <- UploadDao.insert(
                 fixupUploadCreate(dbUser, dbProject, datasource, upload),
-                dbUser
+                dbUser,
+                0
               )
             } yield insertedUpload
 
@@ -80,7 +81,7 @@ class UploadDaoSpec
                 datasource = datasource.id,
                 projectId = Some(dbProject.id)
               )
-              insertedUpload <- UploadDao.insert(uploadToInsert, dbUser)
+              insertedUpload <- UploadDao.insert(uploadToInsert, dbUser, 0)
             } yield (insertedUpload, dbProject)
 
             val (dbUpload, dbProject) =
@@ -130,7 +131,7 @@ class UploadDaoSpec
                 projectId = Some(dbProject.id),
                 layerId = Some(dbLayer.id)
               )
-              insertedUpload <- UploadDao.insert(uploadToInsert, dbUser)
+              insertedUpload <- UploadDao.insert(uploadToInsert, dbUser, 0)
             } yield (insertedUpload, dbProject, dbLayer)
 
             val (dbUpload, dbProject, dbLayer) =
@@ -172,7 +173,8 @@ class UploadDaoSpec
               datasource <- unsafeGetRandomDatasource
               insertedUpload <- UploadDao.insert(
                 fixupUploadCreate(dbUser, dbProject, datasource, insertUpload),
-                dbUser
+                dbUser,
+                0
               )
             } yield
               (insertedUpload, dbUser, dbOrg, dbPlatform, dbProject, datasource)
@@ -196,7 +198,8 @@ class UploadDaoSpec
                   ).toUpload(
                     dbUser,
                     (dbPlatform.id, false),
-                    Some(dbPlatform.id)
+                    Some(dbPlatform.id),
+                    0
                   )
                 UploadDao.update(fixedUpUpdateUpload, uploadId) flatMap {
                   (affectedRows: Int) =>